+++ tools/optimizer/simple_ast.h
@@ -698,7 +698,7 @@ struct JSPrinter {
 
   void printStats(Ref stats) {
     bool first = true;
-    for (int i = 0; i < stats->size(); i++) {
+    for (size_t i = 0; i < stats->size(); i++) {
       Ref curr = stats[i];
       if (!isNothing(curr)) {
         if (first) first = false;
@@ -731,7 +731,7 @@ struct JSPrinter {
     emit(node[1]->getCString());
     emit('(');
     Ref args = node[2];
-    for (int i = 0; i < args->size(); i++) {
+    for (size_t i = 0; i < args->size(); i++) {
       if (i > 0) (pretty ? emit(", ") : emit(','));
       emit(args[i]->getCString());
     }
@@ -993,7 +993,7 @@ struct JSPrinter {
     printChild(node[1], node, 0);
     emit('(');
     Ref args = node[2];
-    for (int i = 0; i < args->size(); i++) {
+    for (size_t i = 0; i < args->size(); i++) {
       if (i > 0) (pretty ? emit(", ") : emit(','));
       printChild(args[i], node, 0);
     }
@@ -1023,7 +1023,7 @@ struct JSPrinter {
     emit('{');
     newline();
     Ref cases = node[2];
-    for (int i = 0; i < cases->size(); i++) {
+    for (size_t i = 0; i < cases->size(); i++) {
       Ref c = cases[i];
       if (!c[0]) {
         emit("default:");
@@ -1057,7 +1057,7 @@ struct JSPrinter {
   void printVar(Ref node) {
     emit("var ");
     Ref args = node[1];
-    for (int i = 0; i < args->size(); i++) {
+    for (size_t i = 0; i < args->size(); i++) {
       if (i > 0) (pretty ? emit(", ") : emit(','));
       emit(args[i][0]->getCString());
       if (args[i]->size() > 1) {
@@ -1165,7 +1165,7 @@ struct JSPrinter {
   void printArray(Ref node) {
     emit('[');
     Ref args = node[1];
-    for (int i = 0; i < args->size(); i++) {
+    for (size_t i = 0; i < args->size(); i++) {
       if (i > 0) (pretty ? emit(", ") : emit(','));
       print(args[i]);
     }
@@ -1177,7 +1177,7 @@ struct JSPrinter {
     indent++;
     newline();
     Ref args = node[1];
-    for (int i = 0; i < args->size(); i++) {
+    for (size_t i = 0; i < args->size(); i++) {
       if (i > 0) {
         pretty ? emit(", ") : emit(',');
         newline();
@@ -1388,7 +1388,7 @@ public:
     assert(switch_[0] == SWITCH);
     assert(code[0] == BLOCK);
     if (!explicitBlock) {
-      for (int i = 0; i < code[1]->size(); i++) {
+      for (size_t i = 0; i < code[1]->size(); i++) {
         switch_[2]->back()->back()->push_back(code[1][i]);
       }
     } else {
